+++ src/pages/ProductDetail.js
@@ -8,16 +8,17 @@ import ProductReview from '../components/ProductDetail/ProductReview';
 import ProductInquiry from '../components/ProductDetail/ProductInquiry';
 import { getProductDetail } from '../api/product';
 import { postHelp, deleteHelp } from '../api/product';
+import { postLike, deleteLike } from '../api/product';
 
 const ProductDetail = () => {
     const [productInfo, setProductInfo] = useState();
-    const { id } = useParams();
+    const productId = useParams().id;
     const memberId = localStorage.getItem('memberId');
 
     // 제품 상세 정보 받아오기
     const getInfo = async () => {
         try {
-            const res = await getProductDetail(id, memberId);
+            const res = await getProductDetail(productId, memberId);
             setProductInfo(res);
         } catch (err) {
             console.log(err);
@@ -28,9 +29,35 @@ const ProductDetail = () => {
         getInfo();
     }, []);
 
-    // 도움돼요 클릭 시
-    const handleHelpClick = async () => {
+    // 찜 관련
+    // 찜하기 & 찜 취소하기
+    const handleLike = async () => {
         try {
+            if (!productInfo?.isLiked) {
+                const res = await postLike(productId, memberId);
+            } else {
+                const res = await deleteLike(productId, memberId);
+            }
+            getInfo();
+        } catch (err) {
+            console.log(err);
+        }
+    };
+
+    // 찜하기
+    const handleHeartClick = () => {
+        handleLike(memberId);
+    };
+
+    // 도움돼요 클릭 시
+    const handleHelpClick = async (reviewId, isReviewHelp) => {
+        try {
+            if (!isReviewHelp) {
+                const res = await postHelp(reviewId, memberId);
+            } else {
+                const res = await deleteHelp(reviewId, memberId);
+            }
+            getInfo();
         } catch (err) {
             console.log(err);
         }
@@ -39,13 +66,17 @@ const ProductDetail = () => {
     return (
         <>
             <Header />
-            <ProductAtf productInfo={productInfo} />
+            <ProductAtf
+                productInfo={productInfo}
+                handleHeartClick={handleHeartClick}
+            />
             <NavBar reviewNum={productInfo?.reviewCount} />
             <ProductDescription productInfoImg={productInfo?.productInfoImg} />
             <ProductReview
                 reviews={productInfo?.reviews}
                 reviewImg={productInfo?.reviewImages}
                 reviewCount={productInfo?.reviewCount}
+                handleHelpClick={handleHelpClick}
             />
             <ProductInquiry inquiries={productInfo?.inquiries} />
         </>
